@@ -22,15 +22,15 @@ keywords = [
   "webm",
 ]
 readme = "README.md"
-requires-python = ">=3.10,<=3.11"
+requires-python = ">=3.12"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
     "Operating System :: OS Independent",
 ]
 dependencies = [
-    "PySide6 ~= 6.4.0",
-    "requests ~= 2.28.1"
+    "pyside6 ~= 6.6.0",
+    "requests ~= 2.31.0"
 ]
 
 [tool.hatch.version]
